+++ src/streamdiffusion/acceleration/tensorrt/engine_pool.py
@@ -10,7 +10,7 @@ import torch
 from polygraphy import cuda
 
 from .controlnet_engine import ControlNetModelEngine, HybridControlNet
-from .controlnet_models import create_controlnet_model
+from .models.controlnet_models import create_controlnet_model
 from .builder import EngineBuilder, create_onnx_path
 from ...model_detection import detect_model
