+++ src/streamdiffusion/acceleration/tensorrt/models/__init__.py
@@ -1,0 +1,13 @@
+from .models import Optimizer, BaseModel, CLIP, UNet, VAE, VAEEncoder
+from .controlnet_models import ControlNetTRT, ControlNetSDXLTRT
+
+__all__ = [
+    "Optimizer",
+    "BaseModel", 
+    "CLIP",
+    "UNet",
+    "VAE",
+    "VAEEncoder",
+    "ControlNetTRT",
+    "ControlNetSDXLTRT",
+]
